+++ src/main/java/net/silentchaos512/gear/block/MetalBlock.java
@@ -1,0 +1,17 @@
+package net.silentchaos512.gear.block;
+
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IWorldReader;
+
+public class MetalBlock extends Block {
+    public MetalBlock(Properties properties) {
+        super(properties);
+    }
+
+    @Override
+    public boolean isBeaconBase(BlockState state, IWorldReader world, BlockPos pos, BlockPos beacon) {
+        return true;
+    }
+}