+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Command to show parts GUI (`/sgear_parts show_gui`). Does not require OP perms. [#107]
+- Crimson iron and steel blocks [#88]
 ### Changed
 - Updated compatible items to remove from drop system of Mine and Slash (AzureDoom) [#97]
 - `/sgear_parts list` no longer requires OP perms